Formatting of files Addition/removal of dependencies Move impl agnostic socket factory to correct module
