+++ junixsocket-adapter/pom.xml
@@ -26,6 +26,11 @@
             <systemPath>${basedir}/src/main/lib/junixsocket-1.3.jar</systemPath>
         </dependency>
 
+        <dependency>
+            <groupId>com.github.nexus</groupId>
+            <artifactId>nexus-socket</artifactId>
+        </dependency>
+
     </dependencies>
 
     <build>
