+++ junixsocket-adapter/src/main/resources/META-INF/services/com.github.nexus.socket.UnixSocketFactory
