+++ nexus-socket/pom.xml
@@ -1,5 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>nexus-socket</artifactId>
@@ -18,26 +21,12 @@
             <groupId>com.github.nexus</groupId>
             <artifactId>nexus-config</artifactId>
         </dependency>
-        
-        <dependency>
-            <groupId>com.github.nexus</groupId>
-            <artifactId>junixsocket-adapter</artifactId>
-        </dependency>
 
         <dependency>
             <groupId>com.github.nexus</groupId>
             <artifactId>nexus-security</artifactId>
         </dependency>
-        
-        <dependency>
-            <groupId>junixsocket</groupId>
-            <artifactId>junixsocket</artifactId>
-            <scope>runtime</scope>
-            <version>1.3</version>
-        </dependency>
 
     </dependencies>
-
-
 
 </project>
