+++ nexus-socket/src/main/java/com/github/nexus/socket/SocketServer.java
@@ -1,6 +1,5 @@
 package com.github.nexus.socket;
 
-import com.github.nexus.junixsocket.adapter.UnixSocketFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
