+++ nexus-socket/src/main/java/com/github/nexus/socket/SocketServerFactory.java
@@ -1,8 +1,6 @@
 package com.github.nexus.socket;
 
-
 import com.github.nexus.config.Config;
-import com.github.nexus.junixsocket.adapter.UnixSocketFactory;
 
 import java.util.concurrent.Executors;
 
@@ -10,13 +8,13 @@ public interface SocketServerFactory {
 
     static SocketServer createSocketServer(final Config config) {
         try {
-        return new SocketServer(
-            config.getUnixSocketFile(),
-            new HttpProxyFactory(config.getServerConfig()),
-            Executors.newCachedThreadPool(),
-            UnixSocketFactory.create()
-        );
-        } catch(Exception ex) {
+            return new SocketServer(
+                config.getUnixSocketFile(),
+                new HttpProxyFactory(config.getServerConfig()),
+                Executors.newCachedThreadPool(),
+                UnixSocketFactory.create()
+            );
+        } catch (final Exception ex) {
             throw new NexusSocketException(ex);
         }
     }
