+++ nexus-socket/src/test/java/com/github/nexus/socket/SocketServerTest.java
@@ -1,6 +1,5 @@
 package com.github.nexus.socket;
 
-import com.github.nexus.junixsocket.adapter.UnixSocketFactory;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
