+++ nexus-socket/src/test/java/com/github/nexus/socket/UnixDomainSocketIT.java
@@ -1,6 +1,5 @@
 package com.github.nexus.socket;
-
-import com.github.nexus.junixsocket.adapter.UnixSocketFactory;
+;
 import com.github.nexus.socket.client.UnixDomainClientSocket;
 import org.junit.Test;
