+++ nexus-socket/src/test/java/com/github/nexus/socket/client/UnixDomainClientSocketTest.java
@@ -1,7 +1,7 @@
 package com.github.nexus.socket.client;
 
-import com.github.nexus.junixsocket.adapter.UnixSocketFactory;
 import com.github.nexus.socket.NexusSocketException;
+import com.github.nexus.socket.UnixSocketFactory;
 import org.assertj.core.api.Assertions;
 import org.junit.After;
 import org.junit.Before;
