+++ nexus-socket/src/test/resources/META-INF/services/com.github.nexus.socket.UnixSocketFactory
