+++ pom.xml
@@ -54,12 +54,22 @@
                     <version>1.4.0</version>
                 </plugin>
 
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-dependency-plugin</artifactId>
+                    <version>3.1.1</version>
+                </plugin>
+
             </plugins>
 
         </pluginManagement>
 
-
         <plugins>
+
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-dependency-plugin</artifactId>
+            </plugin>
 
             <plugin>
                 <groupId>org.pitest</groupId>
@@ -599,7 +609,5 @@
         </dependency>
 
     </dependencies>
-
-
 
 </project>
