+++ rest-server-parent/rest-server/src/main/java/com/github/nexus/server/RestServer.java
@@ -1,9 +1,9 @@
 package com.github.nexus.server;
 
 public interface RestServer {
-        
+
     void start() throws Exception;
-    
+
     void stop() throws Exception;
-  
+
 }
