+++ service-locator-parent/service-locator/src/main/java/com/github/nexus/service/locator/ServiceLocator.java
@@ -6,10 +6,9 @@ import java.util.Set;
 public interface ServiceLocator {
 
     Set<Object> getServices(String filename);
-    
+
     static ServiceLocator create() {
         return ServiceLoader.load(ServiceLocator.class).iterator().next();
     }
-    
-    
+
 }
